+++ sdk/src/client/api/mod.rs
@@ -11,6 +11,7 @@ mod stream_events;
 mod subscribe;
 mod sync;
 mod try_connect;
+mod unsubscribe;
 mod util;
 
 pub use self::add::*;
@@ -26,3 +27,4 @@ pub use self::stream_events::*;
 pub use self::subscribe::*;
 pub use self::sync::*;
 pub use self::try_connect::*;
+pub use self::unsubscribe::*;